# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "JSON/REST based Ham Radio Logging Server",
        "title": "QSOLink",
        "contact": {}
    },
    "paths": {
        "/api/qso": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QSO"
                ],
                "summary": "Get all QSO records",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/qso.Qso"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QSO"
                ],
                "summary": "Create QSO record",
                "parameters": [
                    {
                        "description": "QSO data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/qso.Qso"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/qso.Qso"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/qso/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QSO"
                ],
                "summary": "Get QSO record by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "QSO ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/qso.Qso"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QSO"
                ],
                "summary": "Update QSO record by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "QSO ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "QSO data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/qso.Qso"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/qso.Qso"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QSO"
                ],
                "summary": "Delete QSO record by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "QSO ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "qso.Qso": {
            "type": "object",
            "properties": {
                "band": {
                    "type": "integer",
                    "example": 20
                },
                "callsign": {
                    "type": "string",
                    "example": "W1AW"
                },
                "city": {
                    "type": "string",
                    "example": "New Haven"
                },
                "country": {
                    "type": "string",
                    "example": "USA"
                },
                "county": {
                    "type": "string",
                    "example": "New Haven"
                },
                "createdAt": {
                    "type": "string"
                },
                "dateon": {
                    "type": "string",
                    "example": "2023-10-28"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "mode": {
                    "type": "string",
                    "example": "cw"
                },
                "name": {
                    "type": "string",
                    "example": "John Smith"
                },
                "power": {
                    "type": "integer",
                    "example": 100
                },
                "qslr": {
                    "type": "boolean",
                    "example": false
                },
                "qsls": {
                    "type": "boolean",
                    "example": false
                },
                "remarks": {
                    "type": "string",
                    "example": "This is a test contact with W1AW"
                },
                "rstr": {
                    "type": "integer",
                    "example": 599
                },
                "rsts": {
                    "type": "integer",
                    "example": 599
                },
                "state": {
                    "type": "string",
                    "example": "CT"
                },
                "timeon": {
                    "type": "string",
                    "example": "16:34:58.947Z"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}